fix windows cmake and compiler warnings

+++ src/common/tiglcommonfunctions.cpp
@@ -642,7 +642,7 @@ void GetShapeExtension(const TopoDS_Shape& shape,
 int GetHash(const TopoDS_Shape& shape)
 {
 #if OCC_VERSION_HEX >= VERSION_HEX_CODE(7,8,0)
-    return std::hash<TopoDS_Shape>{}(shape);
+    return static_cast<int>(std::hash<TopoDS_Shape>{}(shape));
 #else
     return shape.HashCode(INT_MAX);
 #endif
